requirements.txt
pydrive2
torchinfo
torch
torchvision
pillow
tqdm
bisect
scipy